Add files via upload
In this exercise we're learning more about strings. Z. Shaw is showing us various ways of making them, combining them, setting them as values for variables and then embedding them in other strings by using those variables, etc. He asks that we then take this exercise and give it to a friend, have that friend break our code, and then it's up to us to repair our code. That will be forthcoming once my friend gets my code broken and sends it back to me to repair.