Add files via upload
As I said in an earlier submission Zed Shaw instructs us to have a friend break our code. From now on when I submit an exercise with "br.py" following the number in the title that means that it is a piece of clean code that I have sent to my mentor Michael so that he can break my code (thus "br" for "break"). I'll submit files titled "ex#brkn.py" with "brkn" standing for "broken" meaning that this is the broken code that Michael has sent me and I'll submit files titled "ex#fxd.py" with "fxd" standing for "fixed" so you can see my solutions to his problems.