Add files via upload
So this is actually starting to get very interesting. In the last exercise we learned how to open a separate text file within a script, then read that file within the script, and then print that file in the terminal. In this exercise we're learning how to edit separate text files (deleting their contents and writing new text) from within our script.