Add files via upload
This is the first exercise that has REALLY stumped me. I'm starting to get a hang of it but I had to really sit for an hour and parse it out. The really confusing thing was trying to figure out how the '.readline()' command was switching from line to line. I finally looked at the FAQ section where Z. Shaw explains that there's code built into that command that automatically causes it to move through the file line by line. I'm going to be looking through this exercise for a bit because I still feel like I have quite a bit to learn from teasing it out further.