Add files via upload
I'm still working on ex21.py. I'm trying to understand it as best I can before I actually run it because I want to try to predict what it will do and then see if I'm correct. This is part of trying to understand the 'return' function and how it works. This little experiment gave me a sense of how return stores the output of the function, within the computer, as the new value attached to the function. My difficulty was in trying to see how to get that stored value to return as an expression within the terminal once the script has ceased to run the function. I figured out how to do that here but I'm wondering if there's a more efficient way to do it. I'm continuing to explore this so more to come later.