Add files via upload
So this is interesting but supper hard. Zed is teaching us about bits, bytes, strings, and how they are all converted, one into the other and back again. I am going to have to read and reread this lesson several times though the general concepts are pretty clear after the first read-through.