Add files via upload
A quiz in which we are supposed to take broken, imperfect, or poorly written code, from another programer, and fix it and clean it up so that it runs better. I'll be submitting my work on this in installments (via branches that I'll integrate into the master branch) as I progress through it.